Conflicts:
	Compiler/src/MainChristian.scala

+++ Compiler/src/exastencils/polyhedron/Extractor.scala
@@ -30,7 +30,7 @@ object Extractor {
   private final val SKIP_ANNOT : String = "PolySkip"
 
   /** set of all functions that are allowed in a scop (these must not have side effects) */
-  private final val allowedFunctions = HashSet[String]("sin", "cos", "tan", "sinh", "cosh", "tanh", "exp", "sqrt")
+  private final val allowedFunctions = Set[String]("abs", "fabs") ++= MathFunctions.signatures.keys
 
   /** set of symbolic constants that must not be modeled as read accesses (these must be constant inside a scop) */
   private final val symbolicConstants = HashSet[String]()
@@ -52,8 +52,8 @@ object Extractor {
 
     expr match {
 
-      case varAcc : VariableAccess =>
-        val islStr : String = ScopNameMapping.expr2id(varAcc)
+      case _ : VariableAccess | _ : ArrayAccess =>
+        val islStr : String = ScopNameMapping.expr2id(expr)
         if (vars != null)
           vars.add(islStr)
         constraints.append(islStr)
@@ -62,12 +62,6 @@ object Extractor {
       case str : StringLiteral if (ScopNameMapping.id2expr(str.value).isDefined) =>
         val e = ScopNameMapping.id2expr(str.value).get
         val islStr : String = ScopNameMapping.expr2id(e)
-        if (vars != null)
-          vars.add(islStr)
-        constraints.append(islStr)
-
-      case array : ArrayAccess =>
-        val islStr : String = ScopNameMapping.expr2id(array)
         if (vars != null)
           vars.add(islStr)
         constraints.append(islStr)
@@ -83,45 +77,50 @@ object Extractor {
         gParConstr.append(" and ")
         constraints.append('(').append(islStr).append("=1)")
 
-      case OffsetIndex(min, max, ind, off) =>
-        constraints.append('(')
-        bool |= extractConstraints(ind, constraints, formatString, lParConstr, gParConstr, vars)
-        constraints.append('+')
-        bool |= extractConstraints(off, constraints, formatString, lParConstr, gParConstr, vars)
-        constraints.append(')')
-        if (lParConstr != null) off match {
-          case _ : VariableAccess | _ : ArrayAccess =>
-            lParConstr.append('(').append(min).append("<=")
-            lParConstr.append(ScopNameMapping.expr2id(off))
-            lParConstr.append("<=").append(max).append(')')
-            lParConstr.append(" and ")
-
-          case MultiplicationExpression(ListBuffer(IntegerConstant(c), arr : ArrayAccess)) =>
-            lParConstr.append('(').append(min).append("<=").append(c).append('*')
-            lParConstr.append(ScopNameMapping.expr2id(arr))
-            lParConstr.append("<=").append(max).append(')')
-            lParConstr.append(" and ")
-
-          case MultiplicationExpression(ListBuffer(arr : ArrayAccess, IntegerConstant(c))) =>
-            lParConstr.append('(').append(min).append("<=").append(c).append('*')
-            lParConstr.append(ScopNameMapping.expr2id(arr))
-            lParConstr.append("<=").append(max).append(')')
-            lParConstr.append(" and ")
-
-          case _ =>
+      case bExpr @ BoundedExpression(min, max, _ : VariableAccess | _ : ArrayAccess) =>
+        val islStr : String = ScopNameMapping.expr2id(bExpr, bExpr.expr)
+        if (vars != null)
+          vars.add(islStr)
+        constraints.append(islStr)
+        if (lParConstr != null) {
+          lParConstr.append('(').append(min).append("<=").append(islStr).append("<=").append(max).append(')')
+          lParConstr.append(" and ")
         }
-      //        val islStr : String = ScopNameMapping.expr2id(expr)
-      //        if (vars != null)
-      //          vars.add(islStr)
-      //        constraints.append(islStr)
-      //        if (paramConstr != null) {
-      //          paramConstr.append(" and ")
-      //          paramConstr.append('(').append(min).append('+')
-      //          extractConstraints(ind, paramConstr, formatString, null, vars)
-      //          paramConstr.append("<=").append(islStr).append("<=")
-      //          extractConstraints(ind, paramConstr, formatString, null, vars)
-      //          paramConstr.append('+').append(max).append(')')
-      //        }
+
+        // case OffsetIndex(min, max, ind, off) =>
+        //   off match {
+        //     case ArrayAccess(_ : iv.IterationOffsetBegin, _, _) =>
+        //       off.annotate(SimplifyExpression.EXTREMA_ANNOT, (min.toLong, max.toLong)) // preserve extrema information since OffsetIndex will be lost
+        //     case ArrayAccess(_ : iv.IterationOffsetEnd, _, _) =>
+        //       off.annotate(SimplifyExpression.EXTREMA_ANNOT, (min.toLong, max.toLong)) // preserve extrema information since OffsetIndex will be lost
+        //     case _ => // nothing to do
+        //   }
+        //   constraints.append('(')
+        //   bool |= extractConstraints(ind, constraints, formatString, lParConstr, gParConstr, vars)
+        //   constraints.append('+')
+        //   bool |= extractConstraints(off, constraints, formatString, lParConstr, gParConstr, vars)
+        //   constraints.append(')')
+        //   if (lParConstr != null) off match {
+        //     case _ : VariableAccess | _ : ArrayAccess =>
+        //       lParConstr.append('(').append(min).append("<=")
+        //       lParConstr.append(ScopNameMapping.expr2id(off))
+        //       lParConstr.append("<=").append(max).append(')')
+        //       lParConstr.append(" and ")
+        //
+        //     case MultiplicationExpression(ListBuffer(IntegerConstant(c), arr : ArrayAccess)) =>
+        //       lParConstr.append('(').append(min).append("<=").append(c).append('*')
+        //       lParConstr.append(ScopNameMapping.expr2id(arr))
+        //       lParConstr.append("<=").append(max).append(')')
+        //       lParConstr.append(" and ")
+        //
+        //     case MultiplicationExpression(ListBuffer(arr : ArrayAccess, IntegerConstant(c))) =>
+        //       lParConstr.append('(').append(min).append("<=").append(c).append('*')
+        //       lParConstr.append(ScopNameMapping.expr2id(arr))
+        //       lParConstr.append("<=").append(max).append(')')
+        //       lParConstr.append(" and ")
+        //
+        //     case _ =>
+        //   }
 
       case iff : iv.IndexFromField =>
         val islStr : String = ScopNameMapping.expr2id(iff)
@@ -258,8 +257,6 @@ object Extractor {
         constraints.append(')')
         bool = true
 
-      // FIXME: support min/max exp
-
       case _ => throw new ExtractionException("unknown expression: " + expr.getClass() + " - " + expr.prettyprint())
     }
 
@@ -283,6 +280,21 @@ object Extractor {
 
     return str
   }
+
+  private def checkCode(name : String) : Unit = {
+
+    var i : Int = name.length()
+    while (i > 0) {
+      i -= 1
+      name.charAt(i) match {
+        case '=' | '+' | '*' | '/' | '[' | '(' | ' ' =>
+          throw new ExtractionException("expression in string constant found: " + name)
+        case '-' if name.charAt(i + 1) != '>' =>
+          throw new ExtractionException("expression in string constant found: " + name)
+        case _ =>
+      }
+    }
+  }
 }
 
 private final case class ExtractionException(msg : String) extends Exception(msg)
@@ -303,6 +315,9 @@ class Extractor extends Collector {
   /** indicates if a new scop starting at the next node can be merged with the last one */
   private var mergeScops : Boolean = false
 
+  /** stack of additional conditions for the next statements found */
+  private val conditions = new ArrayStack[String]()
+
   /** all found static control parts */
   val scops = new ArrayBuffer[Scop](256)
   val trash = new ArrayBuffer[(Node, String)] // TODO: debug; remove
@@ -310,15 +325,8 @@ class Extractor extends Collector {
   private object curScop {
 
     object curStmt {
-
       private var id_ : Int = -1
       private var label_ : String = null
-      private val df = {
-        val df_ = new java.text.DecimalFormat()
-        df_.setMinimumIntegerDigits(4) // 9999 different statements should be enough
-        df_.setGroupingUsed(false)
-        df_
-      }
 
       def exists() : Boolean = { label_ != null }
       def label() : String = { label_ }
@@ -326,7 +334,7 @@ class Extractor extends Collector {
       def leave() : Unit = { label_ = null }
       def next() : this.type = {
         id_ += 1
-        label_ = "S" + df.format(id_)
+        label_ = "S%04d".format(id_)
         return this
       }
     }
@@ -370,12 +378,11 @@ class Extractor extends Collector {
       return origLoopVars_
     }
 
-    // [..] -> { %s[..] : .. }
-    def buildIslSet(tupleName : String) : isl.Set = {
+    // [..] -> { %s[..] : .. %s }
+    def buildIslSet(tupleName : String, cond : String = "") : isl.Set = {
       formatterResult.delete(0, Int.MaxValue)
-      formatter.format(setTemplate_, tupleName)
-      val set = isl.Set.readFromStr(Isl.ctx, formatterResult.toString())
-      return set
+      formatter.format(setTemplate_, tupleName, cond)
+      return isl.Set.readFromStr(Isl.ctx, formatterResult.toString())
     }
 
     // [..] -> { %s[..] -> %s[%s] }
@@ -438,7 +445,7 @@ class Extractor extends Collector {
     try {
       if (!curScop.exists())
         node match {
-          case loop : LoopOverDimensions with PolyhedronAccessable =>
+          case loop : LoopOverDimensions with PolyhedronAccessible =>
             loop.indices.annotate(SKIP_ANNOT)
             loop.stepSize.annotate(SKIP_ANNOT)
             if (loop.condition.isDefined)
@@ -455,6 +462,10 @@ class Extractor extends Collector {
         node match {
 
           // process
+          case c : ConditionStatement =>
+            c.condition.annotate(SKIP_ANNOT)
+            enterCondition(c)
+
           case a : AssignmentStatement =>
             enterAssign(a)
 
@@ -478,16 +489,6 @@ class Extractor extends Collector {
             index.annotate(SKIP_ANNOT)
             enterArrayAccess(varName, index)
 
-          case ArrayAccess(ppVec : iv.PrimitivePositionBegin, index, _) =>
-            ppVec.annotate(SKIP_ANNOT)
-            index.annotate(SKIP_ANNOT)
-            enterArrayAccess(ppVec.prettyprint(), index)
-
-          case ArrayAccess(ppVec : iv.PrimitivePositionEnd, index, _) =>
-            ppVec.annotate(SKIP_ANNOT)
-            index.annotate(SKIP_ANNOT)
-            enterArrayAccess(ppVec.prettyprint(), index)
-
           case ArrayAccess(tmp : iv.TmpBuffer, index, _) =>
             tmp.annotate(SKIP_ANNOT)
             index.annotate(SKIP_ANNOT)
@@ -504,9 +505,34 @@ class Extractor extends Collector {
             extent.annotate(SKIP_ANNOT)
             enterTempBufferAccess(buffer, index)
 
+          case LoopCarriedCSBufferAccess(buffer, index) =>
+            buffer.annotate(SKIP_ANNOT)
+            index.annotate(SKIP_ANNOT)
+            enterLoopCarriedCSBufferAccess(buffer, index)
+
           case d : VariableDeclarationStatement =>
             d.dataType.annotate(SKIP_ANNOT)
             enterDecl(d)
+
+          // for the following 4 matches: do not distinguish between different elements of
+          //    PrimitivePositionBegin and PrimitivePositionEnd (conservative approach)
+          case ArrayAccess(ppVec : iv.PrimitivePositionBegin, index, _) =>
+            ppVec.annotate(SKIP_ANNOT)
+            index.annotate(SKIP_ANNOT)
+            enterScalarAccess(replaceSpecial(ppVec.prettyprint()))
+
+          case ArrayAccess(ppVec : iv.PrimitivePositionEnd, index, _) =>
+            ppVec.annotate(SKIP_ANNOT)
+            index.annotate(SKIP_ANNOT)
+            enterScalarAccess(replaceSpecial(ppVec.prettyprint()))
+
+          case MemberAccess(ppVec : iv.PrimitivePositionBegin, _) =>
+            ppVec.annotate(SKIP_ANNOT)
+            enterScalarAccess(replaceSpecial(ppVec.prettyprint()))
+
+          case MemberAccess(ppVec : iv.PrimitivePositionEnd, _) =>
+            ppVec.annotate(SKIP_ANNOT)
+            enterScalarAccess(replaceSpecial(ppVec.prettyprint()))
 
           // ignore
           case FunctionCallExpression(name, _) if (allowedFunctions.contains(name)) =>
@@ -525,6 +551,9 @@ class Extractor extends Collector {
             | _ : DivisionExpression
             | _ : ModuloExpression
             | _ : PowerExpression
+            | _ : MinimumExpression
+            | _ : MaximumExpression
+            | _ : CommentStatement
             | NullStatement => // nothing to do for all of them...
 
           // deny
@@ -545,24 +574,32 @@ class Extractor extends Collector {
       isWrite = false
     }
 
-    if (node.removeAnnotation(SKIP_ANNOT).isDefined)
+    if (node.removeAnnotation(SKIP_ANNOT).isDefined) {
       skip = false
-
-    node match {
-      case l : LoopOverDimensions           => leaveLoop(l)
-      case _ : AssignmentStatement          => leaveAssign()
-      case _ : StringLiteral                => leaveScalarAccess()
-      case _ : VariableAccess               => leaveScalarAccess()
-      case _ : ArrayAccess                  => leaveArrayAccess()
-      case _ : DirectFieldAccess            => leaveFieldAccess()
-      case _ : TempBufferAccess             => leaveTempBufferAccess()
-      case _ : VariableDeclarationStatement => leaveDecl()
-      case _                                =>
+      return
     }
+    if (skip)
+      return
+
+    if (curScop.exists())
+      node match {
+        case l : LoopOverDimensions           => leaveLoop(l)
+        case c : ConditionStatement           => leaveCondition(c)
+        case _ : AssignmentStatement          => leaveAssign()
+        case _ : StringLiteral                => leaveScalarAccess()
+        case _ : VariableAccess               => leaveScalarAccess()
+        case _ : ArrayAccess                  => leaveArrayAccess()
+        case _ : DirectFieldAccess            => leaveFieldAccess()
+        case _ : TempBufferAccess             => leaveTempBufferAccess()
+        case _ : LoopCarriedCSBufferAccess    => leaveLoopCarriedCSBufferAccess()
+        case _ : VariableDeclarationStatement => leaveDecl()
+        case _                                =>
+      }
   }
 
   override def reset() : Unit = {
     curScop.discard()
+    conditions.clear()
     isRead = false
     isWrite = false
     skip = false
@@ -573,7 +610,7 @@ class Extractor extends Collector {
 
   /////////////////// methods for node processing \\\\\\\\\\\\\\\\\\\
 
-  private def enterLoop(loop : LoopOverDimensions with PolyhedronAccessable, mergeWithPrev : Boolean) : Unit = {
+  private def enterLoop(loop : LoopOverDimensions with PolyhedronAccessible, mergeWithPrev : Boolean) : Unit = {
 
     for (step <- loop.stepSize)
       if (step != IntegerConstant(1))
@@ -655,6 +692,7 @@ class Extractor extends Collector {
     tmp = templateBuilder.length
     templateBuilder.append(':')
     templateBuilder.append(constrs)
+    templateBuilder.append("%s") // additional constraints (e.g. from conditions)
     templateBuilder.append('}')
     val setTemplate : String = templateBuilder.toString()
 
@@ -676,6 +714,20 @@ class Extractor extends Collector {
     }
   }
 
+  private def enterCondition(cond : ConditionStatement) : Unit = {
+    if (cond.falseBody.isEmpty) {
+      val sb = new StringBuilder(" and ")
+      extractConstraints(cond.condition, sb, false)
+      conditions.push(sb.toString())
+    } else
+      throw new ExtractionException("cannot deal with a non-empty falseBody in a ConditionStatement: " + cond.prettyprint())
+  }
+
+  private def leaveCondition(cond : ConditionStatement) : Unit = {
+    if (cond.falseBody.isEmpty)
+      conditions.pop()
+  }
+
   private def enterStmt(stmt : Statement) : Unit = {
 
     val scop : Scop = curScop.get()
@@ -683,7 +735,7 @@ class Extractor extends Collector {
     val label : String = curScop.curStmt.next().label()
     scop.stmts.put(label, (ListBuffer(stmt), curScop.origLoopVars()))
 
-    val domain = curScop.buildIslSet(label)
+    val domain = curScop.buildIslSet(label, conditions.mkString)
     scop.domain = if (scop.domain == null) domain else scop.domain.addSet(domain)
     val schedule = curScop.buildIslMap(label, "", curScop.modelLoopVars() + ',' + curScop.curStmt.id())
     scop.schedule = if (scop.schedule == null) schedule else scop.schedule.addMap(schedule)
@@ -722,17 +774,7 @@ class Extractor extends Collector {
   private def enterScalarAccess(varName : String, deadAfterScop : Boolean = false) : Unit = {
 
     // hack: filter out code
-    var i : Int = varName.length()
-    while (i > 0) {
-      i -= 1
-      varName.charAt(i) match {
-        case '=' | '+' | '*' | '/' | '[' | '(' | ' ' =>
-          throw new ExtractionException("expression in StringConstant found: " + varName)
-        case '-' if varName.charAt(i + 1) != '>' =>
-          throw new ExtractionException("expression in StringConstant found: " + varName)
-        case _ =>
-      }
-    }
+    checkCode(varName)
 
     if (!curScop.curStmt.exists() || (!isRead && !isWrite))
       throw new ExtractionException("misplaced access expression?")
@@ -746,7 +788,7 @@ class Extractor extends Collector {
 
     val scop : Scop = curScop.get()
 
-    var access : isl.Map = curScop.buildIslMap(curScop.curStmt.label(), replaceSpecial(varName), "")
+    val access : isl.Map = curScop.buildIslMap(curScop.curStmt.label(), replaceSpecial(varName), "")
 
     if (isRead)
       scop.reads = if (scop.reads == null) access else scop.reads.addMap(access)
@@ -763,30 +805,38 @@ class Extractor extends Collector {
     // nothing to do here...
   }
 
-  private def enterArrayAccess(name : String, index : Expression) : Unit = {
+  private def enterArrayAccess(name : String, index : Expression, deadAfterScop : Boolean = false) : Unit = {
 
     if (!curScop.curStmt.exists() || (!isRead && !isWrite))
       throw new ExtractionException("misplaced access expression?")
 
+    // hack: check for code in name
+    checkCode(name)
+
     val scop : Scop = curScop.get()
 
-    var bool : Boolean = false
+    var ineq : Boolean = false
     val indB : StringBuilder = new StringBuilder()
     index match {
       case mInd : MultiIndex =>
         for (i <- mInd) {
-          bool |= extractConstraints(i, indB, false)
+          ineq |= extractConstraints(i, indB, false)
           indB.append(',')
         }
-        indB.deleteCharAt(indB.length - 1)
+        if (!mInd.isEmpty)
+          indB.deleteCharAt(indB.length - 1)
       case ind =>
-        bool |= extractConstraints(ind, indB, false)
+        ineq |= extractConstraints(ind, indB, false)
     }
 
-    if (bool)
+    if (ineq)
       throw new ExtractionException("array access contains (in)equalities")
 
-    var access : isl.Map = curScop.buildIslMap(curScop.curStmt.label(), replaceSpecial(name), indB.toString())
+    val access : isl.Map = curScop.buildIslMap(curScop.curStmt.label(), replaceSpecial(name), indB.toString())
+    if (deadAfterScop) {
+      val dead : isl.Set = access.domain()
+      scop.deadAfterScop = if (scop.deadAfterScop == null) dead else scop.deadAfterScop.addSet(dead)
+    }
 
     if (isRead)
       scop.reads = if (scop.reads == null) access else scop.reads.addMap(access)
@@ -807,7 +857,7 @@ class Extractor extends Collector {
       case SlotAccess(_, offset) => name.append('s').append(offset)
       case s                     => name.append(s.prettyprint())
     }
-    enterArrayAccess(name.toString(), index)
+    enterArrayAccess(replaceSpecial(name.toString()), index)
   }
 
   private def leaveFieldAccess() : Unit = {
@@ -824,6 +874,14 @@ class Extractor extends Collector {
   }
 
   private def leaveTempBufferAccess() : Unit = {
+    leaveArrayAccess()
+  }
+
+  private def enterLoopCarriedCSBufferAccess(buffer : iv.LoopCarriedCSBuffer, index : MultiIndex) : Unit = {
+    enterArrayAccess(buffer.resolveName, index, true)
+  }
+
+  private def leaveLoopCarriedCSBufferAccess() : Unit = {
     leaveArrayAccess()
   }
